Update tests for dlpredictor
