--- Processes/dlpredictor/tests/parse_error_rate.py
@@ -1,33 +1,0 @@
-#  Licensed to the Apache Software Foundation (ASF) under one
-#  or more contributor license agreements.  See the NOTICE file
-#  distributed with this work for additional information
-#  regarding copyright ownership.  The ASF licenses this file
-#  to you under the Apache License, Version 2.0 (the
-#  "License"); you may not use this file except in compliance
-#  with the License.  You may obtain a copy of the License at
-
-#  http://www.apache.org/licenses/LICENSE-2.0.html
-
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-import csv
-
-with file('count_dlpredictor.csv') as f:
-    csv_reader = csv.DictReader(f)
-    row_count = -1
-    for row in csv_reader:
-        if row_count >= 0:
-            uckey = row['uckey']
-            num_days = row['num_days']
-            actual = row['actual']
-            predicted = row['predicted']
-            print('{}  {}  {}  {}'.format(uckey, num_days, actual, predicted))
-        row_count += 1
-    print('num rows: {}'.format(row_count))
-
-
-
